Added bold_font to original cells (that are unmutable) and added a getComparedCellsList(int id) function to get the ids of the cells to compare
getComparedCellsList gives the ids of the 20 cells that are in the same row or column or block of the cell whose id is passed in argument. The result is a QList of ints.
To deal with the bold fonts, I added a IsSudokuValuesFixed QList<bool> that specifies false if the cell is originally empty and true otherwise. Thus, there is a public method to access this Qlist from qml : getIsSudokuValuesFixed.