Fixed the Key.onPressed event in qml + add equivalent Key events + added multiple conflict cells
When you type on 1,2,3,4,5,6,7,8,9 or &,é,",',(,-,è,_,ç you can write the number in the selected cell. To erase any mistake, just click on delete or suppr. When you try to enter a value that would be in conflict to another cell's value ; one or more all of these cells in conflict are now colored in red